split up AddressBook.test.jsx
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/367735
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
Product-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
